Move function / optimize import. Minor refactor
